Raspberry_pi_sensor/requirements.txt
paho-mqtt
adafruit-circuitpython-dht # Gebruikt voor de DHT11
